+++ functional-tests/utils/wait/strata.py
@@ -321,19 +321,3 @@ class StrataWaiter(RpcWaiter):
             timeout=timeout,
             error_with=f"Timeout: Checkpoint index did not increment to expected value({idx})",
         )
-
-    def wait_until_tx_included_in_block(self, txid: str):
-        def _query():
-            try:
-                receipt = self.rpc_client.get_transaction_receipt(txid)
-                return receipt
-            except Exception as e:
-                return e
-
-        result = self._wait_until_with_value(
-            _query,
-            lambda result: not isinstance(result, Exception),
-            error_with="Transaction receipt for txid not available",
-            timeout=self.timeout,
-        )
-        return result
